change all to response

+++ version_1/Trading-System/src/main/java/TradingSystem/Server/DataLayer/Services/ProductService.java
@@ -13,6 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 //import com.journaldev.hibernate.util.HibernateUtil;
 
@@ -33,28 +34,45 @@ public class ProductService {
 
     }
 
-    public int AddProductToStore(int storeID, String productName,
+    public Response AddProductToStore(int storeID, String productName,
                                          String category, Double price, int quantity) {
         DataProduct product = new DataProduct(productName, category, price, quantity);
         DataStore store = storeRepository.getOne(storeID);
         product.setStore(store);
         DataProduct dataProduct =  productRepository.saveAndFlush(product);
-        return dataProduct.getProductID();
+        Response response=new Response(false, "Product was added successfully");
+        response.AddProductID(dataProduct.getProductID());
+        return response;
     }
 
-    public void RemoveProduct(int productId) {
+    public Response RemoveProduct(int productId) {
+        Optional<DataProduct> product=productRepository.findById(productId);
+        if(!product.isPresent()){
+            return new Response(true,"Product id doesn't exist");
+        }
         productRepository.removeDataProductByProductID(productId);
+        return new Response("Removed product successfully");
     }
 
-    public List<DataProduct> findDummyProductByStore(Integer storeID){
-        DataStore store = storeRepository.getOne(storeID);
-        return productRepository.findDummyProductByStore(store);
+    public Response findDummyProductByStore(Integer storeID){
+        Optional<DataStore> store = storeRepository.findById(storeID);
+        if(!store.isPresent()){
+           return new Response(true,"Could not found store id");
+        }
+        List<DataProduct> products= productRepository.findDummyProductByStore(store.get());
+        Response response=new Response(false," ");
+        response.AddDBProductsList(products);
+        return response;
     }
 
-    public void setQuantity(Integer productID, int newQuantity){
-        DataProduct product = productRepository.getOne(productID);
-        product.setQuantity(newQuantity);
-        productRepository.saveAndFlush(product);
+    public Response setQuantity(Integer productID, int newQuantity){
+        Optional<DataProduct> product = productRepository.findById(productID);
+        if(!product.isPresent()){
+            return new Response(true,"Could not find product");
+        }
+        product.get().setQuantity(newQuantity);
+        productRepository.saveAndFlush(product.get());
+        return new Response(false," ");
     }
 
 
@@ -62,25 +80,37 @@ public class ProductService {
         productRepository.deleteAll();
     }
 
-    public void editProductDetails(Integer productID, String productName, Double price, String category, Integer quantity) {
-        DataProduct product = productRepository.getOne(productID);
+    public Response editProductDetails(Integer productID, String productName, Double price, String category, Integer quantity) {
+        Optional<DataProduct> product_opt = productRepository.findById(productID);
+        if(!product_opt.isPresent()){
+            return new Response(true,"Could not find product");
+        }
+        DataProduct product=product_opt.get();
         product.setProductName(productName);
         product.setPrice(price);
         product.setCategory(category);
         product.setQuantity(quantity);
         productRepository.saveAndFlush(product);
+        return new Response(false," ");
     }
 
-    public void addCommentToProduct(Integer productID, Integer userID, String comment) {
-        DataProduct product = productRepository.getOne(productID);
-        DataSubscriber subscriber = subscriberRepository.getOne(userID);
-        DataComment comment1 = new DataComment(subscriber, product, comment);
+    public Response addCommentToProduct(Integer productID, Integer userID, String comment) {
+        Optional<DataProduct> product = productRepository.findById(productID);
+        Optional<DataSubscriber> subscriber = subscriberRepository.findById(userID);
+        if(!product.isPresent() || !subscriber.isPresent()){
+            return new Response(true, "Could not find product or user");
+        }
+        DataComment comment1 = new DataComment(subscriber.get(), product.get(), comment);
         commentRepository.saveAndFlush(comment1);
+        return new Response(false," ");
     }
 
 
 
-    public List<DataProduct> findAllByCategoryAndProductNameAndPriceBetween(String name,String category, int min, int max){
-        return productRepository.findAllByCategoryAndProductNameAndPriceBetween(name,category,min,max);
+    public Response findAllByCategoryAndProductNameAndPriceBetween(String name,String category, int min, int max){
+        List<DataProduct> products=productRepository.findAllByCategoryAndProductNameAndPriceBetween(name,category,min,max);
+        Response response=new Response(false, " ");
+        response.AddDBProductsList(products);
+        return response;
     }
 }
